requirements.txt
boto==2.17.0
mrjob==0.4.2
pytest==2.5.2
pytest-cov==1.6
python-dateutil==2.2
tox==1.7.1
flake8==2.2.0
autopep8==1.0.3
